[Modify] filter && input 의 onSubmit 상위로 이동 #146
