+++ app/src/components/filter/Container.tsx
@@ -1,6 +1,5 @@
 import React, { Dispatch, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { useParams, useSearchParams } from 'react-router-dom';
 import { setCheck, setAreaName, setAreaNo } from '../../redux/filter/reducer';
 import { RootState } from '../../redux/root';
 import MarkBox from '../markbox/MarkBox';
@@ -13,13 +12,11 @@ interface type {
 interface Props {
   setShowFilter: Dispatch<React.SetStateAction<boolean>>;
   showFilter: boolean;
-  resetPageInfo: () => void;
+  onSubmit: (e: any, str: string) => void;
 }
 
-function Filter({ setShowFilter, showFilter, resetPageInfo }: Props) {
+function Filter({ setShowFilter, showFilter, onSubmit }: Props) {
   const filterData: any = useSelector((state: RootState) => state.filter.data);
-  const { no } = useParams();
-  const [searchParams, setSearchParams] = useSearchParams();
   const dispatch = useDispatch();
   const [view, setView] = useState<{ [key: number]: boolean }>({ 0: false });
 
@@ -71,99 +68,6 @@ function Filter({ setShowFilter, showFilter, resetPageInfo }: Props) {
     return '모든 가격대';
   };
 
-  const objDataProcessing = (): any => {
-    const changeNull = (filteringValue: boolean | number | string) => {
-      if (
-        filteringValue === false ||
-        filteringValue === 0 ||
-        filteringValue === 1000000
-      ) {
-        return null;
-      } else return filteringValue;
-    };
-
-    return {
-      check: {
-        sort: {
-          1: changeNull(filterData.check.sort[0]),
-          DESC: changeNull(filterData.check.sort[1]),
-          ASC: changeNull(filterData.check.sort[2]),
-        },
-        target: {
-          0: changeNull(filterData.check.target[0]),
-          1: changeNull(filterData.check.target[1]),
-        },
-        date: {
-          7: changeNull(filterData.check.date[0]),
-          30: changeNull(filterData.check.date[1]),
-          60: changeNull(filterData.check.date[2]),
-          0: changeNull(filterData.check.date[3]),
-        },
-        free: { 1: changeNull(filterData.check.free[0]) },
-      },
-      area: {
-        areaNo: changeNull(filterData.area.areaNo),
-      },
-      price: {
-        min: changeNull(filterData.price.min),
-        max: changeNull(filterData.price.max),
-      },
-    };
-  };
-
-  const sortQuery = () => {
-    if (objDataProcessing().check.sort[1] !== null) {
-      return '&popular=1';
-    } else
-      return (
-        '&sort=' +
-        Object.keys(objDataProcessing().check.sort)
-          .map((el, i) => {
-            if (objDataProcessing().check.sort[el] !== null) {
-              return Object.keys(objDataProcessing().check.sort)[i];
-            }
-          })
-          .filter(el => el)[0]
-      );
-  };
-
-  const drawObjKey = (obj: any) => {
-    const value = Object.keys(obj)
-      .map((el, i) => {
-        if (obj[el] !== null) {
-          return Object.keys(obj)[i];
-        }
-      })
-      .filter(el => el)[0];
-    return value ? value : null;
-  };
-
-  const onSubmit = (e: any) => {
-    const titleQuery = () => {
-      return (
-        searchParams.get('title') !== null &&
-        `&title=${searchParams.get('title')}`
-      );
-    };
-
-    const query = `?categoryNo=${no}${titleQuery()}${sortQuery()}&target=${drawObjKey(
-      objDataProcessing().check.target,
-    )}&date=${drawObjKey(objDataProcessing().check.date)}&free=${drawObjKey(
-      objDataProcessing().check.free,
-    )}&min=${
-      objDataProcessing().check.free[1] === null
-        ? objDataProcessing().price.min
-        : null
-    }&max=${
-      objDataProcessing().check.free[1] === null
-        ? objDataProcessing().price.max
-        : null
-    }&areaNo=${objDataProcessing().area.areaNo}`;
-    e.preventDefault();
-
-    setSearchParams(query);
-  };
-
   return (
     <Presenter
       contents={contents}
@@ -176,7 +80,6 @@ function Filter({ setShowFilter, showFilter, resetPageInfo }: Props) {
       onSubmit={onSubmit}
       setShowFilter={setShowFilter}
       showFilter={showFilter}
-      resetPageInfo={resetPageInfo}
     />
   );
 }
