+++ app/src/components/search/Container.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
 import { useParams, useSearchParams } from 'react-router-dom';
 import { RootState } from '../../redux/root';
@@ -30,7 +30,9 @@ function Search(props: Props) {
   const [localValue, setLocalValue] = useState<string[]>(
     JSON.parse(localStorage.getItem('currentSearch') || '[]'),
   );
-
+  const { no } = useParams();
+  const filterData = useSelector((state: RootState) => state.filter.data);
+  const [searchParams, setSearchParams] = useSearchParams();
   const [showFilter, setShowFilter] = useState(false);
   const [dataList, setDataList] = useState<DataList>({
     hotKey: {
@@ -61,19 +63,110 @@ function Search(props: Props) {
     },
   });
 
-  // 쿼리 순서 상관 없음
-  //https://mo-hae.site//boards/filter
-  //?take=12
-  //&page=${페이지숫자}
-  //&categoryNo=${카테고리숫자}
-  //&sort=${'정렬값'} || &popular=1
-  //&target=${'대상'} || null -> ex) 0=해줄래요(!), 1=구할래요(?)
-  //&date=${일수} || null -> ex) 일주일(7), 1개월(30), 3개월(60)
-  //&free=${1} || null
-  //&min=${최소값} || null
-  //&max=${최대값} || null
-  //&areaNo=${지역숫자} || null
-  //&title=${검색값} || null
+  const objDataProcessing = (): any => {
+    const changeNull = (filteringValue: boolean | number | string) => {
+      if (
+        filteringValue === false ||
+        filteringValue === 0 ||
+        filteringValue === 1000000
+      ) {
+        return null;
+      } else return filteringValue;
+    };
+
+    return {
+      check: {
+        sort: {
+          1: changeNull(filterData.check.sort[0]),
+          DESC: changeNull(filterData.check.sort[1]),
+          ASC: changeNull(filterData.check.sort[2]),
+        },
+        target: {
+          0: changeNull(filterData.check.target[0]),
+          1: changeNull(filterData.check.target[1]),
+        },
+        date: {
+          7: changeNull(filterData.check.date[0]),
+          30: changeNull(filterData.check.date[1]),
+          60: changeNull(filterData.check.date[2]),
+          0: changeNull(filterData.check.date[3]),
+        },
+        free: { 1: changeNull(filterData.check.free[0]) },
+      },
+      area: {
+        areaNo: changeNull(filterData.area.areaNo),
+      },
+      price: {
+        min: changeNull(filterData.price.min),
+        max: changeNull(filterData.price.max),
+      },
+    };
+  };
+
+  const sortQuery = () => {
+    if (objDataProcessing().check.sort[1] !== null) {
+      return '&popular=1';
+    } else
+      return (
+        '&sort=' +
+        Object.keys(objDataProcessing().check.sort)
+          .map((el, i) => {
+            if (objDataProcessing().check.sort[el] !== null) {
+              return Object.keys(objDataProcessing().check.sort)[i];
+            }
+          })
+          .filter(el => el)[0]
+      );
+  };
+
+  const drawObjKey = (obj: any) => {
+    const value = Object.keys(obj)
+      .map((el, i) => {
+        if (obj[el] !== null) {
+          return Object.keys(obj)[i];
+        }
+      })
+      .filter(el => el)[0];
+    return value ? value : null;
+  };
+
+  const onSubmit = (e: React.FormEvent<HTMLFormElement>, str: string) => {
+    const query = `?categoryNo=${no}&title=${
+      value ? value : null
+    }${sortQuery()}&target=${drawObjKey(
+      objDataProcessing().check.target,
+    )}&date=${drawObjKey(objDataProcessing().check.date)}&free=${drawObjKey(
+      objDataProcessing().check.free,
+    )}&min=${
+      objDataProcessing().check.free[1] === null
+        ? objDataProcessing().price.min
+        : null
+    }&max=${
+      objDataProcessing().check.free[1] === null
+        ? objDataProcessing().price.max
+        : null
+    }&areaNo=${objDataProcessing().area.areaNo}`;
+    e.preventDefault();
+
+    if (str === 'search') {
+      if (value.length > 1) {
+        setSearchParams(query);
+
+        localStorage.setItem(
+          'currentSearch',
+          JSON.stringify([value, ...localValue]),
+        );
+        setLocalValue(
+          JSON.parse(localStorage.getItem('currentSearch') || '[]'),
+        );
+      } else alert('두 글자 이상');
+
+      setValue('');
+    } else {
+      setSearchParams(query);
+      resetPageInfo();
+    }
+  };
 
   const hotKeyClick = (e: React.MouseEvent) =>
     console.log(
@@ -107,6 +200,7 @@ function Search(props: Props) {
       hotKeyClick={hotKeyClick}
       setLocalValue={setLocalValue}
       resetPageInfo={resetPageInfo}
+      onSubmit={onSubmit}
     />
   );
 }
