+++ app/src/components/search/Presenter.tsx
@@ -21,7 +21,7 @@ interface InputProps extends Props {
   deleteList: (i: number) => void;
   hotKeyClick: (e: React.MouseEvent) => void;
   setLocalValue: Dispatch<React.SetStateAction<string[]>>;
-  resetPageInfo: () => void;
+  onSubmit: (e: any, str: string) => void;
 }
 
 function Presenter(props: InputProps) {
@@ -37,7 +37,7 @@ function Presenter(props: InputProps) {
     value,
     setValue,
     setLocalValue,
-    resetPageInfo,
+    onSubmit,
   } = props;
   const list: string[] = localValue;
 
@@ -263,7 +263,7 @@ function Presenter(props: InputProps) {
         showFilter={showFilter}
         setShowFilter={setShowFilter}
         setLocalValue={setLocalValue}
-        resetPageInfo={resetPageInfo}
+        onSubmit={onSubmit}
       />
     ) : (
       <Input
@@ -273,7 +273,7 @@ function Presenter(props: InputProps) {
         showFilter={showFilter}
         setShowFilter={setShowFilter}
         setLocalValue={setLocalValue}
-        resetPageInfo={resetPageInfo}
+        onSubmit={onSubmit}
       />
     );
 
@@ -299,7 +299,7 @@ function Presenter(props: InputProps) {
         <Filter
           showFilter={showFilter}
           setShowFilter={setShowFilter}
-          resetPageInfo={resetPageInfo}
+          onSubmit={onSubmit}
         />
       )}
     </div>
